+++ test/types/auth-idx.test-d.ts
@@ -33,6 +33,7 @@ expectType<IdxTransaction>(await authClient.idx.authenticate());
 
 const transaction = await authClient.idx.start();
 expectType<IdxContextUIDisplay | undefined>(transaction.context.uiDisplay);
+expectType<boolean | undefined>(transaction.context.currentAuthenticator.value.deviceKnown);
 
 expectType<InteractResponse>(await authClient.idx.interact());
 expectType<IdxResponse>(await authClient.idx.introspect());